Added system dependencies for 3.0.X versions of passenger. Fixed bug that stopped passenger_conf file from being overridden.

+++ lib/crossroads_capistrano/recipes/passenger.rb
@@ -39,41 +39,43 @@ namespace :passenger do
   end
 
   task :install_deps, :roles => :web do
-    yum.install( {:base => %w(curl-devel httpd-devel apr-devel)}, :stable )
+    yum.install( {:base => %w(curl-devel httpd-devel apr-devel openssl-devel zlib-devel e2fsprogs-devel krb5-devel)}, :stable )
     sudo "rm -rf /etc/httpd/conf.d/ssl.conf"
   end
 
   desc "Set up Apache and Passenger config files"
   task :config, :roles => :web do
-    # You can set the following paths from your deploy.rb file, if needed.
-    set :httpd_site_conf_path, "/etc/httpd/sites-enabled/010-#{application}-#{stage}.conf" unless exists?(:httpd_site_conf_path)
-    set :passenger_conf_path, "/etc/httpd/mods-enabled/passenger.conf" unless exists?(:passenger_conf_path)
+    unless (exists?(:no_passenger_conf) && no_passenger_conf)
+      # You can set the following paths from your deploy.rb file, if needed.
+      set :httpd_site_conf_path, "/etc/httpd/sites-enabled/010-#{application}-#{stage}.conf" unless exists?(:httpd_site_conf_path)
+      set :passenger_conf_path, "/etc/httpd/mods-enabled/passenger.conf" unless exists?(:passenger_conf_path)
 
-    if exists?(:rvm_ruby_string)  # Deploying with RVM
-      ruby_root      = "/usr/local/rvm/wrappers/#{rvm_ruby_string}/ruby"
-      passenger_root = "/usr/local/rvm/gems/#{rvm_ruby_string}/gems/passenger-#{passenger_version}"
-    else  # System Ruby
-      ruby_root      = capture("which ruby").strip
-      gem_path       = capture("ruby -r rubygems -e 'p Gem.path.detect{|p|p.include? \"/usr\"}'").strip.gsub('"','')
-      passenger_root = "#{gem_path}/gems/passenger-#{passenger_version}"
+      if exists?(:rvm_ruby_string)  # Deploying with RVM
+        ruby_root      = "/usr/local/rvm/wrappers/#{rvm_ruby_string}/ruby"
+        passenger_root = "/usr/local/rvm/gems/#{rvm_ruby_string}/gems/passenger-#{passenger_version}"
+      else  # System Ruby
+        ruby_root      = capture("which ruby").strip
+        gem_path       = capture("ruby -r rubygems -e 'p Gem.path.detect{|p|p.include? \"/usr\"}'").strip.gsub('"','')
+        passenger_root = "#{gem_path}/gems/passenger-#{passenger_version}"
+      end
+      # httpd conf
+      sudo "cp -f #{release_path}/config/httpd-rails.conf #{httpd_site_conf_path}"
+      sed httpd_site_conf_path, {"DEPLOY_TO"        => deploy_to,
+                                 "IP_ADDR"          => ip_address,
+                                 "SERVER_NAME"      => site_domain_name,
+                                 "SITE_DOMAIN_NAME" => site_domain_name,
+                                 "HTTP_PORT"        => http_port,
+                                 "HTTPS_PORT"       => https_port}
+      # passenger conf
+      sudo "cp -f #{release_path}/config/passenger.conf #{passenger_conf_path}"
+      sed passenger_conf_path,  {"PASSENGER_ROOT"   => passenger_root,
+                                 "RUBY_ROOT"        => ruby_root}
     end
-    # httpd conf
-    sudo "cp -f #{release_path}/config/httpd-rails.conf #{httpd_site_conf_path}"
-    sed httpd_site_conf_path, {"DEPLOY_TO"        => deploy_to,
-                               "IP_ADDR"          => ip_address,
-                               "SERVER_NAME"      => site_domain_name,
-                               "SITE_DOMAIN_NAME" => site_domain_name,
-                               "HTTP_PORT"        => http_port,
-                               "HTTPS_PORT"       => https_port}
-    # passenger conf
-    sudo "cp -f #{release_path}/config/passenger.conf #{passenger_conf_path}"
-    sed passenger_conf_path,  {"PASSENGER_ROOT"   => passenger_root,
-                               "RUBY_ROOT"        => ruby_root}
   end
 end
 
 before "deploy:cold",        "passenger:install"
-after  "deploy:update_code", "passenger:config" unless (exists?(:no_passenger_conf) && no_passenger_conf)
+after "deploy:update_code",  "passenger:config"
 before "deploy:start",       "deploy:apache_permissions"
 before "deploy:restart",     "deploy:apache_permissions"
